Feat: Add Node and LinkedList classes
- The Node class was added to create new nodes.

- The LinkedList class was added, with methods such as:
append: To add a new node at the end.
prepend: To insert a new node at the beginning.
size: To return the size of the list.
getHead: To return the head of the list.
getTail: To return the tail of the list.
at: to return the node at the given index.
toString: To print the nodes in the list.